requirements.txt
streamlit>=1.36.0
supabase>=2.4.0
google-generativeai>=0.5.0
requests>=2.31.0
pandas>=2.0.0
PyMuPDF>=1.23.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
twilio>=8.0.0
anthropic>=0.28.0
